+++ graphql-maven-plugin-logic/src/main/resources/templates/client_jackson_deserialize.vm.java
@@ -35,7 +35,7 @@ public class CustomScalarDeserializer${object.name}  extends AbstractCustomScala
 #elseif (${object.graphQLScalarTypeGetter})
 				${object.graphQLScalarTypeGetter}()
 #else
-			${object.name} : you must define one of graphQLScalarTypeClass, graphQLScalarTypeStaticField or graphQLScalarTypeGetter (in the POM parameters for CustomScalars)
+			${object.javaName} : you must define one of graphQLScalarTypeClass, graphQLScalarTypeStaticField or graphQLScalarTypeGetter (in the POM parameters for CustomScalars)
 #end
 				);
 	}