+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/main/java/org/allGraphQLCases/Queries.java
@@ -3,6 +3,7 @@ package org.allGraphQLCases;
 import java.util.Date;
 import java.util.List;
 
+import org.allGraphQLCases.client.$break;
 import org.allGraphQLCases.client.$extends;
 import org.allGraphQLCases.client.AllFieldCases;
 import org.allGraphQLCases.client.AllFieldCasesInput;
@@ -12,7 +13,6 @@ import org.allGraphQLCases.client.Episode;
 import org.allGraphQLCases.client.FieldParameterInput;
 import org.allGraphQLCases.client.Human;
 import org.allGraphQLCases.client.HumanInput;
-import org.allGraphQLCases.client.MyBreak;
 import org.allGraphQLCases.graphql.DirectQueries;
 import org.allGraphQLCases.graphql.PreparedQueries;
 import org.allGraphQLCases.graphql.WithBuilder;
@@ -65,7 +65,7 @@ public interface Queries {
 	////////////////////////////////////////////////////////////////////////////
 	// Third part: check of GraphQL types that are java keywords
 
-	public MyBreak aBreak($extends test, String $if)
+	public $break aBreak($extends test, String $if)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException;
 
 	////////////////////////////////////////////////////////////////////////////