+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Not released yet
+
+Any mode, client or server:
+- the GraphQL schema may now use java keywords (if the GraphQL schema uses identifier that are java keyword, these identifier are prefixed by an underscore in the generated code)
+- Input parameters are now managed for scalar fields 
+
 
 # 1.3.1